requirements.txt
confluent_kafka
confluent-kafka
delta-spark
pandas
pyarrow
apache-airflow-providers-apache-spark

